Avoid int overflow when squaring the vector length in MaskLength
v.x and v.y are int, so (v.x*v.x + v.y*v.y) was computed in 32-bit
int before being cast to double; for very large pel-scaled motion
vectors (high resolution at pel=4) the squared sum can exceed INT_MAX
and overflow (UB), yielding a garbage mask value. Promote the operands
to double before multiplying.

+++ src/MotionBlockPyramid.cpp
@@ -1756,7 +1756,7 @@ bool MotionBlockPyramid::IsCompatibleForRecalc(const FramePyramid &other) const 
 
 template<typename PixelType>
 static PixelType MaskLength(VECTOR v, uint8_t pel, float fMaskNormFactor2, float fHalfGamma, int maxVal) noexcept {
-    double norme = (double)(v.x * v.x + v.y * v.y) / (pel * pel);
+    double norme = ((double)v.x * v.x + (double)v.y * v.y) / ((double)pel * pel);
 
     double l = maxVal * pow(norme * fMaskNormFactor2, fHalfGamma);
